pyproject.toml
[project]
name = "cultivalab"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "ascii-magic>=2.7.4",
    "bcrypt>=5.0.0",
    "matplotlib>=3.10.8",
    "mkdocs>=1.6.1",
    "mkdocs-mermaid2-plugin>=1.2.3",
    "mkdocstrings[python]>=1.0.3",
    "numpy>=2.4.2",
    "python-dotenv>=1.2.2",
    "questionary>=2.1.1",
    "rich>=14.3.2",
    "ruff>=0.15.1",
    "streamlit>=1.55.0",
    "supabase>=2.28.3",
    "typer>=0.23.1",
]

[dependency-groups]
dev = [
    "mkdocs-material>=9.7.5",
    "mkdocstrings[python]>=1.0.3",
    "mypy>=1.19.1",
    "pre-commit>=4.5.1",
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
    "radon>=6.0.1",
]
